Move server selection to module
