+++ client/index.html
@@ -5,7 +5,7 @@
     <meta content="text/html;charset=utf-8" http-equiv="Content-Type">
     <meta content="utf-8" http-equiv="encoding">
     <link rel="stylesheet" href="./static/style.css" />
-    <script defer src="./src/main.js"></script>
+    <script type="module" src="./src/main.js"></script>
     <meta name="viewport" content="width=device-width, initial-scale=0.8, maximum-scale=0.8, user-scalable=no" />
     <meta name="description" content="The not so latest trend in io games. Conquer as much land as possible and try to become biggest of them all. It's not a bowl." />
     <meta name="keywords" content="splixio, splix.io, online multiplayer, strategy, game, strategy game, io games, agar.io, multiplayer, conquer, conquer land, capture, games, fun, massive multiplayer, web browser games, html5, flash, web game">
